Refactor client / config / redux API into core
Questionable changes:
- ability to pass a story decorator function
- pass clearDecorators to ConfigAPI

React/Vue/Angular working (apparently). RN still broken.
