+++ app/react/src/client/preview/index.js
@@ -4,13 +4,9 @@ import { createStore } from 'redux';
 import addons from '@storybook/addons';
 import createChannel from '@storybook/channel-postmessage';
 import qs from 'qs';
-import { StoryStore } from '@storybook/core/client';
-import ClientApi from './client_api';
-import ConfigApi from './config_api';
-import render from './render';
+import { StoryStore, ClientApi, ConfigApi, Actions, reducer } from '@storybook/core/client';
 import init from './init';
-import { selectStory } from './actions';
-import reducer from './reducer';
+import render from './render';
 
 // check whether we're running on node/browser
 const { navigator } = global;
@@ -27,7 +23,7 @@ if (isBrowser) {
   const queryParams = qs.parse(window.location.search.substring(1));
   const channel = createChannel({ page: 'preview' });
   channel.on('setCurrentStory', data => {
-    reduxStore.dispatch(selectStory(data.kind, data.story));
+    reduxStore.dispatch(Actions.selectStory(data.kind, data.story));
   });
   Object.assign(context, { channel, window, queryParams });
   addons.setChannel(channel);
@@ -35,14 +31,13 @@ if (isBrowser) {
 }
 
 const clientApi = new ClientApi(context);
-const configApi = new ConfigApi(context);
-
-// do exports
 export const storiesOf = clientApi.storiesOf.bind(clientApi);
 export const setAddon = clientApi.setAddon.bind(clientApi);
 export const addDecorator = clientApi.addDecorator.bind(clientApi);
 export const clearDecorators = clientApi.clearDecorators.bind(clientApi);
 export const getStorybook = clientApi.getStorybook.bind(clientApi);
+
+const configApi = new ConfigApi({ clearDecorators, ...context });
 export const configure = configApi.configure.bind(configApi);
 
 // initialize the UI
