+++ app/vue/src/client/preview/index.js
@@ -4,13 +4,9 @@ import { createStore } from 'redux';
 import addons from '@storybook/addons';
 import createChannel from '@storybook/channel-postmessage';
 import qs from 'qs';
-import { StoryStore } from '@storybook/core/client';
-import ClientApi from './client_api';
-import ConfigApi from './config_api';
-import render from './render';
+import { StoryStore, ClientApi, ConfigApi, Actions, reducer } from '@storybook/core/client';
 import init from './init';
-import { selectStory } from './actions';
-import reducer from './reducer';
+import render from './render';
 
 // check whether we're running on node/browser
 const { navigator } = global;
@@ -21,13 +17,31 @@ const isBrowser =
 
 const storyStore = new StoryStore();
 const reduxStore = createStore(reducer);
-const context = { storyStore, reduxStore };
+
+const createWrapperComponent = Target => ({
+  functional: true,
+  render(h, c) {
+    return h(Target, c.data, c.children);
+  },
+});
+const decorateStory = (getStory, decorators) =>
+  decorators.reduce(
+    (decorated, decorator) => context => {
+      const story = () => decorated(context);
+      const decoratedStory = decorator(story, context);
+      decoratedStory.components = decoratedStory.components || {};
+      decoratedStory.components.story = createWrapperComponent(story());
+      return decoratedStory;
+    },
+    getStory
+  );
+const context = { storyStore, reduxStore, decorateStory };
 
 if (isBrowser) {
   const queryParams = qs.parse(window.location.search.substring(1));
   const channel = createChannel({ page: 'preview' });
   channel.on('setCurrentStory', data => {
-    reduxStore.dispatch(selectStory(data.kind, data.story));
+    reduxStore.dispatch(Actions.selectStory(data.kind, data.story));
   });
   Object.assign(context, { channel, window, queryParams });
   addons.setChannel(channel);
@@ -35,7 +49,6 @@ if (isBrowser) {
 }
 
 const clientApi = new ClientApi(context);
-const configApi = new ConfigApi(context);
 
 // do exports
 export const storiesOf = clientApi.storiesOf.bind(clientApi);
@@ -43,6 +56,8 @@ export const setAddon = clientApi.setAddon.bind(clientApi);
 export const addDecorator = clientApi.addDecorator.bind(clientApi);
 export const clearDecorators = clientApi.clearDecorators.bind(clientApi);
 export const getStorybook = clientApi.getStorybook.bind(clientApi);
+
+const configApi = new ConfigApi({ ...context, clearDecorators });
 export const configure = configApi.configure.bind(configApi);
 
 // initialize the UI
