+++ examples/crna-kitchen-sink/package.json
@@ -12,6 +12,7 @@
     "@storybook/channels": "file:../../packs/storybook-channels.tgz",
     "@storybook/channel-postmessage": "file:../../packs/storybook-channel-postmessage.tgz",
     "@storybook/components": "file:../../packs/storybook-components.tgz",
+    "@storybook/core": "file:../../packs/storybook-core.tgz",
     "@storybook/react-native": "file:../../packs/storybook-react-native.tgz",
     "@storybook/ui": "file:../../packs/storybook-ui.tgz",
     "react-native-scripts": "1.1.0",
