+++ lib/components/src/highlight_button.js
@@ -5,7 +5,7 @@ export default glamorous.button(
     border: '1px solid rgba(0, 0, 0, 0)',
     font: 'inherit',
     background: 'none',
-    'box-shadow': 'none',
+    boxShadow: 'none',
     padding: 0,
     ':hover': {
       backgroundColor: 'rgba(0, 0, 0, 0.05)',
