+++ lib/core/package.json
@@ -16,6 +16,9 @@
     "dev": "DEV_BUILD=1 nodemon --watch ./src --exec 'yarn prepare'",
     "prepare": "node ../../scripts/prepare.js"
   },
+  "dependencies": {
+    "global": "^4.3.2"
+  },
   "devDependencies": {
     "babel-cli": "^6.26.0"
   }
