+++ lib/core/src/client/preview/client_api.js
@@ -1,13 +1,20 @@
-/* eslint-disable no-underscore-dangle */
+/* eslint no-underscore-dangle: 0 */
+
+const defaultDecorateStory = (getStory, decorators) =>
+  decorators.reduce(
+    (decorated, decorator) => context => decorator(() => decorated(context), context),
+    getStory
+  );
 
 export default class ClientApi {
-  constructor({ channel, storyStore }) {
+  constructor({ channel, storyStore, decorateStory = defaultDecorateStory }) {
     // channel can be null when running in node
     // always check whether channel is available
     this._channel = channel;
     this._storyStore = storyStore;
     this._addons = {};
     this._globalDecorators = [];
+    this._decorateStory = decorateStory;
   }
 
   setAddon(addon) {
@@ -28,6 +35,13 @@ export default class ClientApi {
   storiesOf(kind, m) {
     if (!kind && typeof kind !== 'string') {
       throw new Error('Invalid or missing kind provided for stories, should be a string');
+    }
+
+    if (!m) {
+      // eslint-disable-next-line no-console
+      console.warn(
+        `Missing 'module' parameter for story with a kind of '${kind}'. It will break your HMR`
+      );
     }
 
     if (m && m.hot) {
@@ -64,13 +78,15 @@ export default class ClientApi {
       // wrap the first decorator and so on.
       const decorators = [...localDecorators, ...this._globalDecorators];
 
-      const getDecoratedStory = decorators.reduce(
-        (decorated, decorator) => context => decorator(() => decorated(context), context),
-        getStory
-      );
+      const fileName = m ? m.filename : null;
 
       // Add the fully decorated getStory function.
-      this._storyStore.addStory(kind, storyName, getDecoratedStory);
+      this._storyStore.addStory(
+        kind,
+        storyName,
+        this._decorateStory(getStory, decorators),
+        fileName
+      );
       return api;
     };
 
@@ -84,11 +100,14 @@ export default class ClientApi {
 
   getStorybook() {
     return this._storyStore.getStoryKinds().map(kind => {
+      const fileName = this._storyStore.getStoryFileName(kind);
+
       const stories = this._storyStore.getStories(kind).map(name => {
         const render = this._storyStore.getStory(kind, name);
         return { name, render };
       });
-      return { kind, stories };
+
+      return { kind, fileName, stories };
     });
   }
 }
