+++ lib/core/src/client/preview/config_api.js
@@ -1,16 +1,16 @@
-/* eslint-disable no-underscore-dangle */
+/* eslint no-underscore-dangle: 0 */
 
 import { location } from 'global';
 import { setInitialStory, setError, clearError } from './actions';
-import { clearDecorators } from './';
 
 export default class ConfigApi {
-  constructor({ channel, storyStore, reduxStore }) {
+  constructor({ channel, storyStore, reduxStore, clearDecorators }) {
     // channel can be null when running in node
     // always check whether channel is available
     this._channel = channel;
     this._storyStore = storyStore;
     this._reduxStore = reduxStore;
+    this._clearDecorators = clearDecorators;
   }
 
   _renderMain(loaders) {
@@ -54,7 +54,7 @@ export default class ConfigApi {
         setTimeout(render);
       });
       module.hot.dispose(() => {
-        clearDecorators();
+        this._clearDecorators();
       });
     }
