+++ lib/core/src/client/preview/index.js
@@ -1,3 +1,7 @@
+import * as Actions from './actions';
+import ClientApi from './client_api';
+import ConfigApi from './config_api';
 import StoryStore from './story_store';
+import reducer from './reducer';
 
-export default { StoryStore };
+export default { Actions, ClientApi, ConfigApi, StoryStore, reducer };
