+++ lib/core/src/client/preview/reducer.js
